AND-15267: Lint Errors - LiveData value assignment nullability mismatch
AND-15267: Lint Errors - NullSafeMutableLiveData: LiveData value
assignment nullability mismatch

+++ app/src/main/java/mega/privacy/android/app/meeting/fragments/InMeetingViewModel.kt
@@ -1128,7 +1128,7 @@ class InMeetingViewModel @Inject constructor(
                                 }
                             }
 
-                            chatIdResult.value = chatId
+                            chatIdResult.value = chatId.toLong()
                         }
                     }
                 }
